Replaced unsigned int by size_t in C string module

+++ include/c/string.h
@@ -5,6 +5,8 @@
 extern "C" {
 #endif
 
+#include "types.h"
+
 /* Branchless conditional negate. */
 #define COND_NEGATE(cond, number) (number) = ((number) ^ -(!cond)) + (!cond)
 
@@ -12,7 +14,7 @@ typedef struct
 {
     const char      *base_table;
     const char      *prefix;
-    unsigned int    prefix_len;
+    size_t    prefix_len;
 }                   base_data;
 
 #define MAKE_BASE_TABLE(name, radix, prefix)   \
@@ -20,13 +22,13 @@ static const base_data  base##name = {(radix), (prefix), sizeof(prefix) - 1}
 
 char                *itoa(int nbr, char *buf, int base);
 long int            strtol(const char *ptr, char **endptr, int base);
-unsigned int        strlen(const char *str);
-void                *memset(void *s, int c, unsigned int n);
-void                *memcpy(void *dest, const void *src, unsigned int n);
+size_t              strlen(const char *str);
+void                *memset(void *s, int c, size_t n);
+void                *memcpy(void *dest, const void *src, size_t n);
 char                *strcpy(char *dest, const char *src);
-char                *strncpy(char *dest, const char *src, unsigned int n);
+char                *strncpy(char *dest, const char *src, size_t n);
 char                *strcat(char *dest, const char *src);
-char                *strncat(char *dest, const char *src, unsigned int n);
+char                *strncat(char *dest, const char *src, size_t n);
 
 #ifdef __cplusplus
 }
